refactoring into Model class
