+++ test/test_regressor_pipeline.py
@@ -7,7 +7,7 @@ from sklearn.exceptions import NotFittedError
 
 os.environ["SAM_ML_LOG_LEVEL"] = "debug"
 from sam_ml.models import Pipeline
-from sam_ml.models.main_regressor import SMAC_INSTALLED
+from sam_ml.models.main_model import SMAC_INSTALLED
 from sam_ml.models.regressor import BYR, DTR, EN, ETR, LLCV, RFR, SGDR, XGBR
